reduce duplicate code and return sensible hlmode information for cursors at lineendings
svn path=/trunk/KDE/kdelibs/kate/; revision=1090231

+++ part/document/katetextline.h
@@ -213,7 +213,7 @@ class KateTextLine : public KShared
      * @param pos position of attribute requested
      * @return value of attribute
      */
-    inline uchar attribute (int pos) const
+    inline int attribute (int pos) const
     {
       for (int i=0; i < m_attributesList.size(); i+=3)
       {
